WWW/photo_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
python-multipart
httpx
pytest